requirements.txt
matplotlib~=3.8.0
moviepy~=1.0.3
numpy~=1.26.4
pillow~=9.4.0
scipy~=1.12.0
hsh~=1.1.0
torchvision~=0.15.2
opencv-python~=4.9.0.80